fixup: nits, benchmarkers in separate file
